requirements.txt
beautifulsoup4==4.6.0
coverage==4.4.1
decorator==4.0.11
hupper==1.0
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
Jinja2==2.9.6
Mako==1.0.6
MarkupSafe==1.0
PasteDeploy==1.5.2
pexpect==4.2.1
pickleshare==0.7.4
pluggy==0.4.0
prompt-toolkit==1.0.14
ptyprocess==0.5.1
py==1.4.34
Pygments==2.2.0
pyramid==1.8.4
pyramid-debugtoolbar==4.1
pyramid-ipython==0.2
pyramid-jinja2==2.7
pyramid-mako==1.0.2
pyramid-tm==2.1
pytest==3.1.2
pytest-cov==2.5.1
repoze.lru==0.6
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.10
tox==2.7.0
traitlets==4.3.2
transaction==2.1.2
translationstring==1.3
venusian==1.1.0
virtualenv==15.1.0
waitress==1.0.2
wcwidth==0.1.7
WebOb==1.7.2
WebTest==2.0.27
zope.deprecation==4.2.0
zope.interface==4.4.1
zope.sqlalchemy==0.7.7
